[tools][server][tests] Split sc-server setup from main and test
