+++ sc-tools/sc-server/CMakeLists.txt
@@ -4,20 +4,16 @@ file(GLOB_RECURSE SOURCES "src/*.cpp" "src/*.h" "src/*.hpp")
 
 list(REMOVE_ITEM SOURCES
     "${SC_SERVER_ROOT}/src/main.cpp"
-    "${SC_SERVER_ROOT}/src/sc_server_factory.cpp"
-    "${SC_SERVER_ROOT}/src/sc_server_factory.hpp"
 )
 
 add_executable(sc-server
     "${CMAKE_CURRENT_SOURCE_DIR}/src/main.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/sc_server_factory.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/sc_server_factory.hpp"
 )
 
 add_library(sc-server-lib SHARED ${SOURCES})
 include_directories(${SC_MEMORY_SRC} ${SC_CONFIG_UTILS_SRC} ${GLIB2_INCLUDE_DIRS})
-target_link_libraries(sc-server-lib sc-memory)
-target_link_libraries(sc-server sc-server-lib sc-config-utils)
+target_link_libraries(sc-server-lib sc-memory sc-config-utils)
+target_link_libraries(sc-server sc-server-lib)
 
 if(${SC_CLANG_FORMAT_CODE})
     target_clangformat_setup(sc-server-lib)
