+++ sc-tools/sc-server/tests/sc-server-test.ini
@@ -10,6 +10,8 @@ log_type = Console
 log_file =
 log_level = Debug
 
+repo_path = sc-server-test-repo
+
 [sc-server]
 host = 127.0.0.1
 port = 1028
